+++ config/tests.php
@@ -2,7 +2,11 @@
 
 declare(strict_types=1);
 
+use Psr\EventDispatcher\EventDispatcherInterface;
+use Psr\EventDispatcher\ListenerProviderInterface;
 use Yiisoft\Aliases\Aliases;
+use Yiisoft\EventDispatcher\Dispatcher\Dispatcher;
+use Yiisoft\EventDispatcher\Provider\Provider;
 
 return [
     \Yiisoft\Yii\Gii\GiiInterface::class => new \Yiisoft\Yii\Gii\Factory\GiiFactory(
@@ -18,4 +22,6 @@ return [
             '@root' => dirname(__DIR__) . '/tests/runtime',
         ]
     ),
+    EventDispatcherInterface::class => Dispatcher::class,
+    ListenerProviderInterface::class => Provider::class,
 ];
